# control-tower/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "basePath": "/",
    "paths": {
        "/clients": {
            "get": {
                "description": "get clients by creator with filters.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CLIENT"
                ],
                "summary": "get clients by creator with filters.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Phone number to filter by",
                        "name": "phone",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "email to filter by",
                        "name": "email",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "First Name to filter by",
                        "name": "firstName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Last Name to filter by",
                        "name": "lastName",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "page number. Zero Indexed",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "max number of records.",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.FiltersResponseDto"
                        }
                    }
                }
            },
            "post": {
                "description": "create client.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CLIENT"
                ],
                "summary": "create client.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "create client dto",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.CreateClient"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.ClientDto"
                        }
                    }
                }
            }
        },
        "/clients/{createdBy}/{id}": {
            "get": {
                "description": "mark client as not available for notifications.",
                "tags": [
                    "CLIENT"
                ],
                "summary": "mark client as not available for notifications.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Client ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Client ID",
                        "name": "createdBy",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "301": {
                        "description": "Moved Permanently"
                    }
                }
            }
        },
        "/clients/{id}": {
            "get": {
                "description": "Get client by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CLIENT"
                ],
                "summary": "Get client by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Client ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.ClientDto"
                        }
                    }
                }
            },
            "delete": {
                "description": "delete client by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CLIENT"
                ],
                "summary": "delete client by ID.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Client ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.ClientDto"
                        }
                    }
                }
            },
            "patch": {
                "description": "update client.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CLIENT"
                ],
                "summary": "update client.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "patch client dto",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.PatchClient"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.ClientDto"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "service.ClientDto": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "lastUpdateAt": {
                    "type": "string"
                },
                "optIn": {
                    "type": "boolean"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "service.CreateClient": {
            "type": "object",
            "required": [
                "firstName",
                "lastName"
            ],
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 2
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "lastUpdateAt": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "service.FiltersResponseDto": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.ClientDto"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "service.PatchClient": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 2
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string",
                    "minLength": 1
                },
                "lastName": {
                    "type": "string",
                    "minLength": 1
                },
                "optIn": {
                    "type": "boolean"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
...
    }
}